Minor fixes following the discussion on pull request 224: - intern.c: watch out for left shift overflow - extern.c: refactor caml_output_value_to_block
git-svn-id: http://caml.inria.fr/svn/ocaml/branches/big-marshal@16522 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ byterun/intern.c
@@ -96,7 +96,7 @@ static inline int16_t read16s(void)
 static inline uint32_t read32u(void)
 {
   uint32_t res =
-    (intern_src[0] << 24) + (intern_src[1] << 16)
+    ((uint32_t)(intern_src[0]) << 24) + (intern_src[1] << 16)
     + (intern_src[2] << 8) + intern_src[3];
   intern_src += 4;
   return res;
@@ -105,7 +105,7 @@ static inline uint32_t read32u(void)
 static inline int32_t read32s(void)
 {
   int32_t res =
-    (intern_src[0] << 24) + (intern_src[1] << 16)
+    ((uint32_t)(intern_src[0]) << 24) + (intern_src[1] << 16)
     + (intern_src[2] << 8) + intern_src[3];
   intern_src += 4;
   return res;
